Project/app/requirements.txt
Flask==2.1.0
google-cloud-datastore==2.7.1
google-auth==1.31.0
requests==2.26.0
functions-framework==3.2.0
pymongo
dnspython